* p4a-linear:
  Remove a few warning spotted by gcc-4.6.
  Update boolean type to be 32/64 bit agnostic.

+++ packages/PIPS/linear/src/arithmetique/boolean.h
@@ -25,11 +25,13 @@
 #ifndef BOOLEAN_INCLUDED
 #define BOOLEAN_INCLUDED
 
+#include <stdint.h>
+
 #ifdef bool
     #error newgen header not compatible with stdbool.h and linear header have to be compatible with newgen
 #endif
 
-typedef int bool; /* we cannot use an enum or stdbool because we need to be compatible with newgen */
+typedef intptr_t bool; /* we cannot use an enum or stdbool because we need to be compatible with newgen */
 
 #define false 0
 #define true 1
